Handle relay errors better
Previously, if there was an error during relay, the current pigeon would
attest the message and then never touch it again.  This means that for
the message to get enough attestations, many pigeons had to attempt the
relay.  This is slow, inefficient, and no longer works with the new way
of dividing up messages.

Now, if a relayer hits an error message, it will send that error data
back to Paloma and other pigeons can simply attest to the error message.
This removes the need to attempt the relay by lots of pigeons to clear
out a bad message.

Other changes in this commit:
* Fixed some logging calls that were missing contextual fields
* Split up the queue get requests to behave differently for attesting vs relaying