Speed up pigeon tick and message time
Ticking every 10 seconds for faster processing of new messages

Remove the sleep.  After some testing, it appears that previous changes
have fixed the account sequence mismatch errors.  Pigeon still needs the
locks in place so that it doesn't run over itself, but as long as one
message is in flight at a time, there are no more errors with account
sequence mismatch

One caveat to this.  If another actor uses the validator key, we will
still see account sequence mismatch errors.  The validator key needs to
belong solely to Pigeon