Continue checking if a pigeon is staking
Previously, this was only checked at startup.  The problem is, pigeons
get jailed all the time while they're still running.  The check is never
re-checked and the pigeon keeps trying to send messages to Paloma that
will just be rejected.

Now, we check in with Paloma every 5 seconds to make sure our pigeon is
still staking.  If it's not staking, it doesn't bother trying to do work
that will just be rejected.

The one exception to this is the keep_alive.  We so aggressively jail
pigeons that I don't want anything delaying a keep_alive message.  We
can fire them off and let them fail when jailed.