Fix pigeon consensus calculation
Pigeon has been testing the validator power against a hard coded number
instead of looking for 2/3 consensus of the power of the valset.  This
is incorrect and can result in Pigeon incorrectly believing consensus is
achieved or not.

If consensus is not achieved, pigeon will end up sending messages that
will fail.  If consensus is achieved but pigeon believes otherwise,
Pigeon won't send messages that would succeed.

Now, we're comparing the signed power vs the valset power.  This should
be the correct way of checking consensus.