Add support for Arbitrum mainnet (#253)
Arbitrum, like Optimism, has made some additional changes built on top of the Ethereum stack, and therefore relies on a modified version of the go-ethereum package, commonly referred to as geth. When we started to support Optimism, we were able to simply exchange the original package for the one used by Optimism. This however will not work again, as there is no support for Arbitrum from the Optimism go-ethereum package, as well as the other way around.

This compatibility issue between go-ethereum versions was likely not anticipated when our EVM wrapper package was originally written, as there is little to no type abstraction and references to the package are all over the code base.

Since we're planning to support non EVM chains in the future as well, there will need to be a considerable refactoring, creating an abstraction layer over the entire chain interaction package. During this work, we might want to consider updating the way I have chosen for this change.

With the work above in mind, I have opted to

isolate the minimum needed change set to support Arbitrum
for this given change set, redirect logic calls to the go-ethereum package provided by Arbitrum
we solve the duplicate module name issue by using a fork of the Arbitrum go-ethereum package which has been updated to use different import paths
we solve the duplicate symbol definition of C files during compiling by removing all native C code from the Arbitrum go-ethereum packge fork, and instead pointing consuming code to the original go-ethereum package.
So this change really introduces a somewhat hacky, but most straight forward way for Arbitrum chain support at this point in time. I would have liked to build a full abstraction layer over this instead of manual logic branching, but the work needed for this is simply out of scope for this issue. It will be done once we move to non EVM chains.