Better logic for checking for executed messages
I'm not actually sure how this hasn't been broken for a much longer
time.  When checking if a logic call was already executed, the filter
topic is returning other logic calls with different IDs.  The current
check of "are they any?" isn't sufficient.

Now, we are going through the returned logs and comparing the message ID
of the log to the current message ID