fix: missing interface registrations (#322)
We have a gravity batch stuck that is being relayed over and over again every 10000 blocks. Pigeon actually see the event on the target chain, but fail to inform Paloma because the gravity messages hadn't been registered against the sdk.Msg interface.

This was most likely just forgotten in the bulk of work for gravity. What bothers me is that this happened before with the AddStatusUpdate message, and apparently for some nodes this works just fine, while for others it doesn't and leads to a crash. Definitely needs more understanding. Either the type is registered or it isn't, and if it isn't, it should fail deterministically across all nodes. But it doesn't.

Tested in private test net, this resolves the issue and Pigeon successfully reports the batch as relayed. With this change, we can create a new Pigeon release. Once the majority of validators upgrade, the batch should be removed successfully.